fix: rename feedback and banner service
