+++ banner/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'review-banner'
+rootProject.name = 'banner'
