+++ banner/src/main/java/com/fanda/banner/dto/ReviewResponseDto.java
@@ -1,4 +1,4 @@
-package com.fanda.feedback.dto;
+package com.fanda.banner.dto;
 
 import java.time.LocalDateTime;
