--- banner/src/main/java/com/fanda/review_banner/ReviewBannerApplication.java
@@ -1,13 +1,0 @@
-package com.fanda.review_banner;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ReviewBannerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ReviewBannerApplication.class, args);
-	}
-
-}
