+++ feedback/build.gradle
@@ -23,18 +23,6 @@ repositories {
 	mavenCentral()
 }
 
-ext {
-	set('springCloudVersion', "2023.0.6")
-}
-
-sourceSets {
-	main{
-		resources{
-			srcDir 'src/main/resources'
-		}
-	}
-}
-
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
@@ -44,22 +32,6 @@ dependencies {
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework.kafka:spring-kafka-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
-	runtimeOnly 'com.mysql:mysql-connector-j'
-	implementation platform("software.amazon.awssdk:bom:2.25.4")
-	implementation 'software.amazon.awssdk:bedrockruntime'
-	implementation 'com.fasterxml.jackson.core:jackson-databind'
-	// pdf
-	implementation 'org.apache.pdfbox:pdfbox:2.0.30'
-	// aws sdk for s3
-	implementation platform('software.amazon.awssdk:bom:2.25.4')
-	implementation 'software.amazon.awssdk:s3'
-}
-
-dependencyManagement {
-	imports {
-		mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-	}
 }
 
 tasks.named('test') {
