+++ feedback/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'feedback-improve'
+rootProject.name = 'feedback'
