+++ feedback/src/main/java/com/fanda/feedback/FeedbackApplication.java
@@ -2,10 +2,8 @@ package com.fanda.feedback;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
-@EnableFeignClients(basePackages = "com.fanda.feedback.repository")
 public class FeedbackApplication {
 
 	public static void main(String[] args) {
