+++ feedback/src/main/resources/application.yml
@@ -1,22 +1,6 @@
 spring:
-  config:
-    import: application-secret.yml
   application:
     name: feedback
-  datasource:
-    url: jdbc:mysql://${DB_HOST}:3306/feedback_db?serverTimezone=Asia/Seoul&useUnicode=true&characterEncoding=UTF-8
-    username: ${DB_USERNAME}
-    password: ${DB_PASSWORD}
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  jpa:
-    hibernate:
-      ddl-auto: update
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
-  cloud:
-    compatibility-verifier:
-      enabled: false
+
 server:
-  port: 8004
+  port: 8003
